sort: Permit percentages as values for buffer size flag
Addresses issue #3500, which calls for sort to permit percentage values
for the buffer-size flag. Portability is an issue here, but I don't
think this change constrains `sort`'s portability any more than its
dependencies already do.

To calculate a fraction of the system's memory, naturally we need to
retrieve the system's total memory. GNU's equivalent -- see
`physmem_total` in `lib/physmem.c` in gnulib -- has a series of `ifdef`s
to decide how to retrive the value. Each platform has its own
considerations, and the result is a pretty serious mess.

This patch omits that mess. In exchange, it will only work on platforms
with a functioning libc. The thing is, `sort` already includes libc,
which is why I said this patch doesn't introduce new constraints on
`sort`'s portability.